sql: properly mark sequences as `bigint` in information_schema.sequences
Release note (bug fix): CockroachDB now properly reports `bigint`
in information_schema.sequences.type, for compatibility with PostgreSQL.

+++ pkg/sql/information_schema.go
@@ -857,7 +857,7 @@ CREATE TABLE information_schema.sequences (
 					tree.NewDString(db.GetName()),    // catalog
 					tree.NewDString(scName),          // schema
 					tree.NewDString(table.GetName()), // name
-					tree.NewDString("integer"),       // type
+					tree.NewDString("bigint"),        // type
 					tree.NewDInt(64),                 // numeric precision
 					tree.NewDInt(2),                  // numeric precision radix
 					tree.NewDInt(0),                  // numeric scale
